server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["yusuf ashour <yachour31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
python-decouple = "^3.8"
django-cors-headers = "^4.2.0"
pytest-django = "^4.5.2"
psycopg2-binary = "^2.9.7"
djangorestframework-simplejwt = "^5.3.0"
celery = {extras = ["redis"], version = "^5.3.4"}
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
django-debug-toolbar = "^4.2.0"
requests = "^2.31.0"
google-api-python-client = "^2.99.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
stripe = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "apis.settings"
python_files = ["tests/test_*.py"]